iiassistant/requirements.txt
asgiref==3.7.2
audioread==3.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
decorator==5.1.1
dj-database-url==2.1.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
django-leaflet==0.29.0
djangorestframework==3.14.0
ffmpeg==1.4
ffprobe==0.5
filelock==3.13.4
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.6
install==1.3.5
Jinja2==3.1.3
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2
llvmlite==0.42.0
MarkupSafe==2.1.5
mbutil==0.3.0
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numba==0.59.1
packaging==24.0
platformdirs==4.2.1
pooch==1.8.1
psycopg2==2.9.9
PyAudio==0.2.11
pycparser==2.22
pydub==0.25.1
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.4.4
srt==3.5.3
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.2
transformers==4.40.1
typing_extensions==4.9.0
urllib3==2.1.0
vosk==0.3.44
websockets==12.0
